type fix in onReceive callback

+++ libraries/Wire/Wire.cpp
@@ -51,7 +51,7 @@ uint8_t TwoWire::txBufferLength = 0;
 
 uint8_t TwoWire::transmitting = 0;
 void (*TwoWire::user_onRequest)(void);
-void (*TwoWire::user_onReceive)(int);
+void (*TwoWire::user_onReceive)(size_t);
 
 static int default_sda_pin = SDA;
 static int default_scl_pin = SCL;
@@ -265,7 +265,7 @@ void TwoWire::onRequestService(void)
 	user_onRequest();
 }
 
-void TwoWire::onReceive( void (*function)(int) ) {
+void TwoWire::onReceive( void (*function)(size_t) ) {
   user_onReceive = function;
 }
